chore: update to nextauth v5 beta
