+++ package.json
@@ -10,8 +10,8 @@
     "prisma": "npx prisma migrate dev --name init"
   },
   "dependencies": {
+    "@auth/prisma-adapter": "^2.10.0",
     "@hookform/resolvers": "^5.2.1",
-    "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/extension-accelerate": "^2.0.2",
     "bcrypt": "^6.0.0",
     "next": "15.4.1",
