+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,2 @@
-import { authOptions } from "@/src/lib/auth";
-import NextAuth from "next-auth";
-
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };
+import { handlers } from "@/auth";
+export const { GET, POST } = handlers;
