+++ src/app/dashboard/new-item/page.tsx
@@ -1,11 +1,10 @@
-import { getServerSession } from "next-auth";
-import { authOptions } from "@/src/lib/auth";
 import { redirect } from "next/navigation";
 import NewItem from "../../../components/newItem";
 import Dashboard from "../page";
+import { auth } from "@/auth";
 
 export default async function NewItemPage() {
-  const session = await getServerSession(authOptions);
+  const session = await auth();
 
   if (!session) {
     redirect("/");
