+++ src/app/page.tsx
@@ -26,6 +26,7 @@ export default function Login() {
         const res = await signIn("credentials", {
           ...data,
           redirect: false,
+          callbackUrl: "/dashboard/items", // where to go after login
         });
 
         if (!res?.ok) {
